requirements.txt
# Core web crawling and data extraction
crawl4ai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Financial data APIs
yfinance>=0.2.0
alpha-vantage>=2.3.0
fredapi>=0.5.0
quandl>=3.7.0

# SEC filings and financial data
sec-edgar-api>=1.0.0
edgar>=5.0.0

# News and RSS feeds
feedparser>=6.0.0
newspaper3k>=0.2.8

# Data storage and processing
pandas>=2.0.0
numpy>=1.24.0
sqlite3  # Built-in
sqlalchemy>=2.0.0

# Scheduling and monitoring
schedule>=1.2.0
APScheduler>=3.10.0

# Alerts and notifications
requests>=2.31.0
smtplib  # Built-in
twilio>=8.0.0

# Web scraping utilities
selenium>=4.15.0
requests-html>=0.10.0

# Data visualization (optional)
plotly>=5.17.0
dash>=2.14.0

# Environment and configuration
python-decouple>=3.8